+++ .github/workflows/cd.yaml
@@ -17,6 +17,40 @@ env:
   VITE_MATTRAX_CLOUD_ORIGIN: https://cloud.mattrax.app
 
 jobs:
+  sst:
+    name: SST
+    runs-on: ubuntu-latest
+    environment: production
+    concurrency:
+      group: production
+    # if: github.ref == 'refs/heads/main' || github.event_name == 'workflow_dispatch' # TODO
+    steps:
+      - name: Git clone the repository
+        uses: actions/checkout@v4
+
+      - name: Configure AWS credentials
+        uses: aws-actions/configure-aws-credentials@v4
+        with:
+          role-to-assume: "arn:aws:iam::101829795063:role/mattrax-gh-actions"
+          aws-region: us-east-1
+          role-session-name: mattrax-sst
+
+      - uses: pnpm/action-setup@v4
+        with:
+          version: latest
+
+      - name: Install SST
+        run: curl -fsSL https://ion.sst.dev/install | bash
+
+      - run: sst deploy --stage prod
+        # env:
+        # ARM_USE_OIDC: true
+        # ARM_CLIENT_ID: a17b56f1-0b10-4029-9a89-7f703d3573f8
+        # ARM_TENANT_ID: 22d6679c-fc23-425a-b69b-e5e604dd80db
+        # AZURE_SUBSCRIPTION_ID: 22d6679c-fc23-425a-b69b-e5e604dd80db
+        # OAUTH_CLIENT_ID: kXdvzkEgiN11CNTRL
+        # OAUTH_CLIENT_SECRET: ${{ secrets.TAILSCALE_OAUTH_SECRET }}
+
   build-mattrax:
     name: Build Mattrax
     runs-on: ubuntu-latest
@@ -56,44 +90,6 @@ jobs:
 
           wrangler r2 object put "static/mattrax/$HASH/x86_64-unknown-linux" --file=target/x86_64-unknown-linux-musl/release/mattrax --cache-control "public, max-age=31536000, immutable"
           echo "$HASH" | wrangler r2 object put "static/nightly" --pipe
-
-  # sst:
-  #   name: SST
-  #   runs-on: ubuntu-latest
-  #   # This is required to workaround the lack of wildcard for OIDC scope
-  #   # https://github.com/Azure/azure-workload-identity/issues/373
-  #   #
-  #   # I swear to god Microsoft have never tried anything they have built.
-  #   environment: production
-  #   concurrency:
-  #     group: production
-  #   if: github.ref == 'refs/heads/main' || github.event_name == 'workflow_dispatch'
-  #   steps:
-  #     - name: Git clone the repository
-  #       uses: actions/checkout@v4
-
-  #     - name: Configure AWS credentials
-  #       uses: aws-actions/configure-aws-credentials@v4
-  #       with:
-  #         role-to-assume: "arn:aws:iam::101829795063:role/mattrax-gh-actions"
-  #         aws-region: us-east-1
-  #         role-session-name: mattrax-sst-workflow
-
-  #     - uses: pnpm/action-setup@v4
-  #       with:
-  #         version: latest
-
-  #     - name: Install SST
-  #       run: curl -fsSL https://ion.sst.dev/install | bash
-
-  #     - run: cd infra && sst deploy --stage brendonovich
-  #       env:
-  #         ARM_USE_OIDC: true
-  #         ARM_CLIENT_ID: a17b56f1-0b10-4029-9a89-7f703d3573f8
-  #         ARM_TENANT_ID: 22d6679c-fc23-425a-b69b-e5e604dd80db
-  #         AZURE_SUBSCRIPTION_ID: 22d6679c-fc23-425a-b69b-e5e604dd80db
-  #         OAUTH_CLIENT_ID: kXdvzkEgiN11CNTRL
-  #         OAUTH_CLIENT_SECRET: ${{ secrets.TAILSCALE_OAUTH_SECRET }}
 
   build-js:
     strategy:
