+++ apps/landing/package.json
@@ -5,7 +5,6 @@
 	"scripts": {
 		"dev": "vinxi dev --host 0.0.0.0 --port 3001",
 		"build": "vinxi build",
-		"cbuild": "NITRO_PRESET=cloudflare_pages vinxi build",
 		"start": "vinxi start"
 	},
 	"dependencies": {
