--- infra/cloud/Caddyfile
@@ -1,43 +1,0 @@
-{
-	storage mysql
-	admin off
-	email hello@mattrax.app
-	on_demand_tls {
-		ask http://localhost:9000/internal/caddy?auth={$INTERNAL_SECRET}
-	}
-}
-
-mattrax.fly.dev {
-	redir https://mdm.mattrax.app
-	header X-Region {$FLY_REGION}
-	header Server Mattrax
-}
-
-mdm.mattrax.app {
-	tls {
-		client_auth {
-			mode verify_if_given
-			trust_pool file {
-				pem_file /mtls-roots.pem
-			}
-		}
-	}
-
-	reverse_proxy localhost:9000
-	header X-Region {$FLY_REGION}
-}
-
-enterpriseenrollment.mattrax.app {
-	reverse_proxy localhost:9000
-	header X-Region {$FLY_REGION}
-}
-
-https:// {
-	tls {
-		on_demand
-	}
-
-	reverse_proxy localhost:9000
-	header X-Region {$FLY_REGION}
-	header Server Mattrax
-}
